fix: resolve dependency cycle
